requirements.txt
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
blackboxprotobuf==1.0.1
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
cryptography==42.0.3
debugpy==1.8.0
decorator==5.1.1
executing==2.0.1
Flask==3.0.1
Flask-Cors==4.0.0
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
ipykernel==6.29.0
ipython==8.12.3
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.0
jupyter_core==5.7.1
lxml==5.1.0
lz4==4.3.3
MarkupSafe==2.1.4
matplotlib-inline==0.1.6
more-itertools==10.2.0
nest-asyncio==1.6.0
numpy==1.24.4
packaging==23.2
pandas==2.0.3
parso==0.8.3
pendulum==3.0.0
pickleshare==0.7.5
pillow==10.2.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
protobuf==3.10.0
psutil==5.9.8
pure-eval==0.2.2
pyahocorasick==2.0.0
PyAudio==0.2.14
pycparser==2.21
pycryptodomex==3.20.0
Pygments==2.17.2
PyJWT==2.8.0
Pymem==1.13.1
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32==306
pywxdump==2.4.35
pyzmq==25.1.2
requests==2.31.0
requests-file==2.0.0
requests-toolbelt==1.0.0
silk-python==0.2.4
simple-salesforce==1.12.5
six==1.16.0
stack-data==0.6.3
time-machine==2.13.0
tornado==6.4
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
wcwidth==0.2.13
Werkzeug==3.0.1
zeep==4.2.1
zipp==3.17.0
